clean up after merging with refactor
