+++ owiddata/03.integrateDataSources.py
@@ -25,7 +25,8 @@ def pair_datasource_names(viper_table, owid_names):
                                      orient="index",
                                      columns=owid_names)
 
-    # Identify the best hit for each VIPER and each OWID vax name
+    # Identify the best hit for each OWID vax name, since these are usually
+    # a subset of the VIPER names
     owid_bestmatch = heatMap.idxmax(axis=0).to_dict() # row max
     viper_bestmatch = heatMap.idxmax(axis=1).to_dict()
     print(owid_bestmatch)
